Added labels to LinkButtons in about page.
Note: while not having labels does have a cleaner look, having labels on
the buttons allows the user to hover over them and preview the URLs
before they click on them. Also, prevents that 'ghost button' highlight
when the action row is clicked and the button is activated -- it makes
more sense for it to highlight with a label.