Removed duplicate shuffle controls.
Reason: It makes more sense to have to have the shuffle mode option in
only one place (the Settings page), than to have it appear twice. Plus,
it allows the Gallery page to shine as a place to view and manage the
user's individual Bing backgrounds, without the shuffle controls kind of
randomly mishmashed with it.